condensed identity and useoktaauth in one component
