+++ dotcom-rendering/src/lib/identity.tsx
@@ -1,6 +1,12 @@
 import { getCookie } from '@guardian/libs';
-import type { AuthState, OktaAuthOptions } from '@okta/okta-auth-js';
-import { OktaAuth } from '@okta/okta-auth-js';
+import type { AuthState, OktaAuthOptions } from '@okta/own-lib';
+import { OktaAuth } from '@okta/own-lib';
+import React, { useEffect, useState } from 'react';
+
+interface OktaAuthContextType {
+	oktaAuth: OktaAuth;
+	authState: AuthState | null;
+}
 
 const CLIENT_ID = '0oa53x6k5wGYXOGzm0x7';
 const ISSUER =
@@ -16,9 +22,9 @@ const config: OktaAuthOptions = {
 	pkce: true,
 };
 
-export const oktaAuthInit = new OktaAuth(config);
+const oktaAuthInit = new OktaAuth(config);
 
-export async function isLoggedIn(oktaAuth: OktaAuth, authState: AuthState) {
+async function isLoggedIn(oktaAuth: OktaAuth, authState: AuthState) {
 	console.log('using isLoggedIn function');
 	if (ifSignedOutClearTokens(oktaAuth)) {
 		console.log('signed out and cleared tokens');
@@ -27,7 +33,7 @@ export async function isLoggedIn(oktaAuth: OktaAuth, authState: AuthState) {
 	return await checkIfSignedIn(authState, oktaAuth);
 }
 
-export function ifSignedOutClearTokens(oktaAuth: OktaAuth) {
+function ifSignedOutClearTokens(oktaAuth: OktaAuth) {
 	if (getCookie({ name: 'GU_SO' })) {
 		oktaAuth.tokenManager.clear();
 		return true;
@@ -35,10 +41,7 @@ export function ifSignedOutClearTokens(oktaAuth: OktaAuth) {
 	return false;
 }
 
-export async function checkIfSignedIn(
-	authState: AuthState,
-	oktaAuth: OktaAuth,
-) {
+async function checkIfSignedIn(authState: AuthState, oktaAuth: OktaAuth) {
 	if (authState.isAuthenticated) {
 		console.log('is authenticated with tokens');
 		return true;
@@ -59,3 +62,57 @@ export async function checkIfSignedIn(
 	}
 	return false;
 }
+
+const useOktaAuth = (): OktaAuthContextType => {
+	const oktaAuth = oktaAuthInit;
+	const [authState, setAuthState] = React.useState<AuthState | null>(() => {
+		return oktaAuth.authStateManager.getAuthState();
+	});
+
+	React.useEffect(() => {
+		async function oktaAuthState() {
+			// Update Security provider with latest authState
+			const currentAuthState = oktaAuth.authStateManager.getAuthState();
+			if (currentAuthState !== authState) {
+				setAuthState(currentAuthState);
+			}
+			const handler = (newAuthState: AuthState) => {
+				setAuthState(newAuthState);
+			};
+			oktaAuth.authStateManager.subscribe(handler);
+
+			// Trigger an initial change event to make sure authState is latest
+			await oktaAuth.start();
+
+			return () => {
+				oktaAuth.authStateManager.unsubscribe(handler);
+			};
+		}
+		oktaAuthState().catch(() => {
+			console.log('error in oktaAuthState');
+		});
+	}, [oktaAuth, authState]);
+	return { oktaAuth, authState };
+};
+
+export const CheckUserSignInStatus = (): boolean => {
+	const { oktaAuth, authState } = useOktaAuth();
+	console.log(authState);
+	console.log(oktaAuth);
+
+	const [isSignedIn, setIsSignedIn] = useState(false);
+	useEffect(() => {
+		(async () => {
+			if (authState) {
+				setIsSignedIn(await isLoggedIn(oktaAuth, authState));
+			}
+		})().catch(() => {
+			console.log('error in oktaAuthState');
+		});
+	}, [authState, oktaAuth]);
+
+	console.log('checkSignedIn');
+	console.log(isSignedIn);
+
+	return isSignedIn ? true : false;
+};
